MWPFE-37-moved homepage code to HomePage component
